dbcsv/requirements.txt
fastapi==0.115.12
pydantic==2.11.1
uvicorn==0.34.0
uvicorn[standard]
pydantic-settings==2.0.0
python-dotenv
PyJWT==2.10.1
python-multipart
PyYAML
passlib[bcrypt]>=1.7.4
pytest
requests
lark
